+++ apps/publish/src/make/externals/timetree.org/fetchField.ts
@@ -1,0 +1,15 @@
+import axios from "axios"
+const arrayBufferToString = (buffer: ArrayBufferLike | ArrayLike<number>) =>
+    String.fromCharCode.apply(null, Array.from(new Uint8Array(buffer).values()))
+const fetchField = async (url: string): Promise<string | null> => {
+    try {
+        const response = await axios.get(url, { responseType: "arraybuffer" })
+        const text = arrayBufferToString(response.data)
+        const [, value] = text.split(":", 2)
+        return value.trim() || null
+    } catch (e) {
+        console.warn(e)
+        throw e
+    }
+}
+export default fetchField