+++ packages/utils/CHANGELOG.md
@@ -9,7 +9,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+-   New functions: `extractUUIDv4()` and `isNonnegativeNumber()`.
+
 ### Changed
+
+-   `parseIdentifier()` now decodes URI comppnents.
 
 ### Deprecated